FiledTest/requirements.txt
mysqlclient
django==3.1.2
djangorestframework